Toggle
Have added a new Toggle to the main window. This was done due to the previous one being toggled whenever the user clicked anywhere within the cell.

Have actually used the Toggle and have also customised the appearance to kick it back to the checkmark toggle rather than the default green blob