MapView removed + DetailView/NewEntryView + Textfield in NewEntryView
The MapView has been commented out completely for the time being and I doubt that I'll add it back in but I'll keep it around just incase

The naming of the DetailView and NewEntryView has been sorted. So where the user enters their item is now the NewEntryView and if they want to view their item: they'll head to the DetailView

Have ported over the Textfield for entering the name of the item and save functionality. Currenlty doesn't accept input but I assume that's something to do with either a Boolean value or the TextLimit that I've placed on it