ContentView Display + TextFields and TextLimits on them
Have altered how each of the items are displayed within the cells on the ContentView. The alignment of the VStack now ensures that the name of the item and quantity are nicely aligned ontop of one another.

Have changed to TextFields to allow more flexibility and especailly with the extra notes section that needed to be mutliline. The amount that the user can input has also been increased