ContentView (Quantity)
Now have the quantity being displayed on the ContentView. The issue lay with explicitly stating that it was Int16 I was using both in the Extension of the model and in the NewEntryView.

Have changed the default amount of an item as zero is daft so the Stepper now starts at 1