Trimming (.whitespacesAndNewlines) + NavigationButtons
Have managed to implement the whitespacesAndNewlines for when the user is entering the item that they're after. If the user were to continually mash enter, then type the item, and then save it: the ContentView would display the whitespace.

The Nav buttons were clear on the Detail and NewEntry view so changed that init to ensure that they all show as white