DetailView - Adjusted
Still playing around with how the item and the relevant notes are displayed on the DetailView.

Have moved the item's name away from the .inline navbar and instead throwing it on the body of the view. This was done to ensure that all of the text is disaplayed should the user wish to user a couple of new lines.