TextLimit count + NewEntryView (Picker)
Have implemented a little text limit counter on each of the TextFields just to give the user an idea of the amount of space they have left

Have taken a new direction with the NewEntryView and have opted to get a Picker to determine the measurement of the item they need (liters, kgs, etc) and then letting them specify the numerical amount after that.

Currentlly only issue is getting what is selected from the Picker to appear in the ContentView. This does appear to be how the selected is viewed (should be a String rather than an Int). I will play around with the Picker so that the value that is returned is the desired string rather than an index place on an array (I think it's currently that)