DarkMode toggle
Now have the DarkMode toggling properly and turns out I was just trying to reference the color incorrectly. Now happily displays the blue/black when needed

Tried some digging with getting the keyboard to not cover the TextField but it appears the .ignoresSafeArea(."whatever") makes no difference.

Did insert a custom class that really only works on the button but does nothing for the TextField