Keyboard Bump Up
Have now managed to get the keyboard to bump up when the user goes into the final TextField when they're entering a new note for one of the items on the NewEntryView. Had to put in a helper class that uses NotificationCentre when the keyboard is detected on the UI and then moves the view accordingly.

Have also reduced the size of the note section and changed all of the TextEditors > TextFields as it made the UI look awful on the iPhone Mini. So has happily removed the issue with some of the text not appearing on the DetailView when excessive amount of empty lines were used

May put the button back on the bottom of the UI as it looks a bit naked currently but will toy around with it shortly