Singular/Plural + Picker apperance
Now have the ContentView displaying an s or no s depending on the amount that the user has input. It's not pretty and wouldn't work with something like that has a plural form that goes outwith just adding an 's' so I've been a bit cheeky and removed the outliers for the time being. I guess it could be done with a switch statement in the future.

Have also altered the appearance of the list view for the picker. Would be nice to be able to change a bit more other than the font colour but will keep poking around to see what happens