Picker Options + Trailing NavBarButton
Have added some more options to the Picker to ensure that people have a bit more choice and the trailing NavBarButton has been given a set frame size to ensure that it works across all devices. Was previously an issue where on the iPhone8 the icon was being clipped but fine on anything else.

All good now

Now to look into text recognition