SpriteKit Game
Trying to get the SpriteKit game going on the DetailView when there are no notes. Currently have some some basic game that I've taken from HackingWithSwift  that displays some squares when a user taps on the screen

Have added the Scene to the DetailView but having an issue displaying the game. Will look more into this next