New Picker Options + Marketing
Have added the new Picker options to give the user some more options
Have included the marketing images too for the App Store just so that they're kept somewhere for safekeeping

Have also removed the GameScene code that I was playing around with. I'll come back to it at some point in the future.