onMove (almost working) + Asset Tidy
Have the .onMove working on the List to a degree. The items can now be moved without an issue being thrown about the array being out of bounds. This was solved by ensuring that the array is now a new array that mirrors what is held on CoreData and is then mapped accordingly. Last issue is updating the order correctly but that will be done next

Also some Asset tidyup as there were some old icons that no longer serve a purpose and have therefore been removed