Rename + New Entrys For User + Most Recent App Submit + UnitTesting
Have finally renamed it from ShoppingList > Yer Messages due to playing around with the Unit Testing.

Have included the new entries for the user to specify the type of item + the font has been changed to as it was an issue with this update (but not the last one).

Still waiting to hear back from the Review team to see what their verdict is exactly so not sure what's happening there. The person was running it as an iPad app whilst I've specified in the Build Settings that it's only for iPhone. The app runs fine on the iPad simulator as that horrible iPhone app within an iPad.

Have the basic framework down for the CoreData unit test but need to work on how the CoreData model is referenced I believe before I can test it properly.