App Store Submission + LaunchScreen
Found that issue lay with the ".navigationBarTitleDisplayMode(.inline)" needing to be moved down a bracket to ensure that it works on iPads. So bumped it down and the app has been submitted again for review.

LaunchScreen was being a pain but now have a plain light/dark background and will look to have a proper loading icon for the next time