Background Issue
So I've stopped using the ZStacks as that seems to throw bugs for some other people but never threw anything for me. So instead I've worked with using the init() to set the TableView background (as the TableView is what underpins the List and Form) and also the NavBar. From there I've then used .background(Color("defaultBackground")) accordingly and there doesn't appear to be any issues either in the sim or on my device here