Removal Last ZStack + DetailView + NewEntryView (Save Button)
Found one last ZStack on the ContentView and have changed the modifiers around to ensure that app performs as it shoudl

DetailView now displays the specified item in a better way with the two bodies of text now sitting further up on the view and looking less collapsed. A placeholder now appears shoudl the user not have added any "Extra Notes".

Have also tidied up the button on the NewEntry view. Have dropped the border and just left a shadow off the button instead as it looks a little nicer