Background
It seems a little excessive to have .background(etc) everywhere when the ZStack seemed to have worked perfectly previously but I've added a few more in across the views.

What is being displayed on the sims changes across devices and then across the physical devices so it's hard to get a gauge of things sometimes but I understand SwiftUI is still being developed. Let's hope this update goes through